Install template
Copied files from homepage static website and stored them in the correct
static directories. Updated the home view to render index.html instead.

Updated the links in index.html to load from static and image sources.
Removed navbar section to navbar.html for use in all templates.

Updated the routes in the javascript for the css so it would account for
the static directory

Update the posts page to load data programatically, and created some
dummy test data in a dictionary to test the template.