Add database models
Created the database models in models.py: 1 for posts and 1 for tags.

Linked the models for access in the admin panel.
Ran migrations and created a superuser account to access the admin.

Created some posts in the admin panel and used those to load into the
templates rather than the dictionary. Add tags to the template and added
some css styling.

Fix the links to link to specific posts and load that data from the
database.