Use slug fields instead of id's in urls
Cleaning up the urls to display slugs of the titles instead of the post
ids. Add the slugfield to the Post model and migrate.

Overide the save method, if there is no slug, add one, and ensure that
it's unique, adding a number if it isn't.

Update urls.py, views.py, and all the templates so that the post.id is
replaced by the post.slug parameter.