Add thumbnail field to database
Added a url field to hold an image's url so that it can be displayed on
the post. Edited templates to show images from each post, and added some
styles to the display of the post page.

Still need to add some css rules for the text to display properly.

Changed back to bootstrap 4 to accomodate the crispy forms.